chore(module): extract plugins to separate files
Separate the md `comp` import plugin template to avoid later broken linting
