+++ lib/module.js
@@ -109,13 +109,26 @@ export default function ContentModule(moduleOpts) {
 		}
 	])
 
-	this.addPlugin({
-		src: resolve(__dirname, 'plugin.js'),
-		options: {
-			...options,
-			srcDirFromPlugin: join('~/', srcDir)
-		}
-	})
+  // 4. Add import md comps plugin
+
+  this.addPlugin({
+    src: resolve(__dirname, 'plugins/importMdComps.template.js'),
+    options: {
+      srcDirFromPlugin: join('~/', srcDir)
+    }
+  })
+
+  // 4. Add content plugins
+   
+  this.addPlugin({
+    src: resolve(__dirname, 'plugins/requestContent.js')
+  })
+
+  // 5. Add nuxtent-body component plugin
+
+  this.addPlugin({
+    src: resolve(__dirname, 'plugins/nuxtentBodyComponent.js')
+  })
 }
 
 export { pkg as meta }
