+++ lib/plugins/requestContent.js
@@ -1,46 +1,29 @@
-import Vue from 'vue'
-
 const { join } = require('path')
 
-const options = <%= JSON.stringify(options) %>
-const mdComps = {}
+function toQuery (options = {}) {
+  let exclude = options.exclude
+  if (!exclude) return ''
 
-function importAllMdComps (r) {
-  r.keys().forEach(key => mdComps[key] = r(key))
+  if (Array.isArray(exclude)) return 'exclude=' + exclude.join(',')
+
+  return 'exclude=' + exclude
 }
 
-importAllMdComps(require.context(
-  <%= JSON.stringify(options.srcDirFromPlugin) %>, true, /\.comp\.md$/
-))
-
-Vue.component('nuxtent-body', {
-  functional: true,
-  props: {
-    body: { required: true }
-  },
-  render (h, ctx) {
-   const { body } = ctx.props
-   if (typeof body === 'object') {
-      const MarkdownComponent = mdComps[body.relativePath]
-      return ( <MarkdownComponent /> )
-    } else {
-      return ( <div domPropsInnerHTML={ body } /> )
-    }
-  }
-})
-
-export default ({ app, isServer, isClient, hotReload, route }) => {
-  const isNotContentReq = hotReload || route.fullPath.includes('__webpack_hmr?') || route.fullPath.includes('.hot-update.')
+export default ({ app, isServer, isClient, isStatic, hotReload, route }) => {
+  const isNotContentReq =
+    hotReload ||
+    route.fullPath.includes('__webpack_hmr?') ||
+    route.fullPath.includes('.hot-update.')
   if (isNotContentReq) return
 
-  const isAPI = isServer || !options.isStatic
+  const isAPI = isServer || !isStatic
   const cache = {}
 
   async function fetchContent (path, permalink, query = '') {
     if (isAPI) {
       const apiEndpoint = join(path, permalink + query)
-      if (options.isDev || !cache[apiEndpoint]) {
-         cache[apiEndpoint] = (await app.$axios.get(apiEndpoint)).data
+      if (!isStatic || !cache[apiEndpoint]) {
+        cache[apiEndpoint] = (await app.$axios.get(apiEndpoint)).data
       }
       return cache[apiEndpoint]
     } else if (isClient) {
@@ -51,21 +34,21 @@ export default ({ app, isServer, isClient, hotReload, route }) => {
         cache[browserPath] = (await app.$axios.get(browserPath)).data
       }
       return cache[browserPath]
-    }
-    else {
-      return // static server build
+    } else {
+       // static server build
     }
   }
 
   app.$content = function requestMethod (contentDir) {
     let query = ''
     return {
-      query (options) { // per page query
+      query (options = {}) {
+        // per page query
         query = toQuery(options)
         return this
       },
       async get (permalink) {
-        if (typeof permalink !== 'string') throw Error(`Permalink must be a string.`)
+        if (typeof permalink !== 'string') { throw Error(`Permalink must be a string.`) }
         return await fetchContent(contentDir, permalink, '?' + query)
       },
       async getBetween (permalink, num1or2, num2 = '') {
@@ -86,13 +69,4 @@ export default ({ app, isServer, isClient, hotReload, route }) => {
       }
     }
   }
-}
-
-function toQuery (options = {}) {
-  let exclude = options.exclude
-  if (!exclude) return ''
-
-  if (Array.isArray(exclude)) return 'exclude=' + exclude.join(',')
-
-  return 'exclude=' + exclude
 }
